+++ siprix_voip_sdk_ios/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.25
+- Added new method 'UpgradeToVideo'
+- Fixed handling case when call started with video on hold
+- iOS/Android: Fixed handling device rotation on capturing video
+- iOS/Android: Added new method 'SwitchCamera' 
+  //2025.09.29
+
 ## 1.0.24
 - Fixed remote hold detection (when all media lines are sendonly or inactive)
 - Modified handling 'sips' scheme (put 'sips' to Contact header)